pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Django base template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.8",
    "django>=5.2,<6",
    "django-admin-honeypot>=1.1",
    "django-admin-rangefilter>=0.10",
    "django-admin-sortable2>=2.2",
    "django-allauth>=65.0",
    "django-auditlog>=2.3",
    "django-constance[database]>=3.0",
    "django-debug-toolbar>=4.0",
    "django-environ>=0.12",
    "django-filer>=3.0",
    "django-filter>=24.0",
    "django-imagekit>=5.0",
    "django-import-export>=4.0",
    "django-log-viewer>=1.1",
    "django-modeltranslation>=0.20",
    "django-object-actions>=4.0",
    "django-rosetta>=0.10.3",
    "django-solo>=2.0",
    "django-split-settings>=1.3",
    "django-celery-beat>=2.5",
    "django-celery-results>=2.5",
    "django-summernote==0.8.20.0",
    "drf-spectacular>=0.28",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.5",
    "easy-thumbnails>=2.8",
    "Pillow",
    "sqlparse>=0.5",
    "django-lucus>=0.1.0",
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = ["E", "F", "UP"]
